backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
av==12.2.0
beautifulsoup4==4.12.3
browser-cookie3==0.19.1
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
intel-openmp==2021.4.0
Jinja2==3.1.3
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.1.5
langsmith==0.1.83
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
protobuf==4.25.3
psutil==6.0.0
pycryptodomex==3.20.0
pydantic==2.8.1
pydantic_core==2.20.1
Pygments==2.18.0
pyktok==0.0.22
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.12
tbb==2021.11.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
